Extract shared BottomSheetModal from Changelog & ImageCredit modals
Both modals were the same bottom-sheet shell: a rounded surface sliding up
from the bottom with a drag handle, a title + close-button header, swipe/
backdrop dismissal, and a safe-area bottom inset. Factor that into a
reusable BottomSheetModal that takes a title and children; each modal now
only supplies its own body (and a maxHeight for the changelog).
