+++ src/components/popups/BottomSheetModal.tsx
@@ -1,0 +1,104 @@
+import type { ReactNode } from "react";
+import type { StyleProp, ViewStyle } from "react-native";
+import { StyleSheet, View } from "react-native";
+import Modal from "react-native-modal";
+import { useSafeAreaInsets } from "react-native-safe-area-context";
+
+import { CloseIcon } from "#/components/Icons";
+import Heading from "#/components/typography/Heading";
+import UiPressable from "#/components/ui/UiPressable";
+import UiSpace from "#/components/ui/UiSpace";
+import Colors from "#/constants/Colors";
+import { useAppColorScheme } from "#/hooks/useAppColorScheme";
+
+interface BottomSheetModalProperties {
+  isVisible: boolean;
+  onClose: () => void;
+  title: string;
+  children: ReactNode;
+  /** Extra styles for the sheet container, e.g. a maxHeight. */
+  containerStyle?: StyleProp<ViewStyle>;
+}
+
+/**
+ * A bottom-anchored sheet modal: a rounded surface that slides up from the
+ * bottom with a drag handle, a title + close-button header, and swipe/backdrop
+ * dismissal. Content below the header is provided via children. The bottom
+ * padding includes the safe-area inset so content clears the Android nav bar.
+ */
+const BottomSheetModal = ({
+  isVisible,
+  onClose,
+  title,
+  children,
+  containerStyle,
+}: BottomSheetModalProperties) => {
+  const colorScheme = useAppColorScheme();
+  const insets = useSafeAreaInsets();
+  const corporate = Colors[colorScheme].primary;
+  const surface = Colors[colorScheme].surface;
+  const textMuted = Colors[colorScheme].textMuted;
+
+  return (
+    <Modal
+      isVisible={isVisible}
+      onSwipeComplete={onClose}
+      onBackdropPress={onClose}
+      swipeDirection="down"
+      style={styles.modal}
+    >
+      <View
+        style={[
+          styles.container,
+          { backgroundColor: surface, paddingBottom: 20 + insets.bottom },
+          containerStyle,
+        ]}
+      >
+        <View style={[styles.handle, { backgroundColor: textMuted }]} />
+        <UiSpace size={16} />
+        <View style={styles.header}>
+          <Heading style={styles.title}>{title}</Heading>
+          <UiPressable
+            accessibilityRole="button"
+            accessibilityLabel="Schließen"
+            onPress={onClose}
+            hitSlop={8}
+          >
+            <CloseIcon size={28} color={corporate} />
+          </UiPressable>
+        </View>
+        {children}
+      </View>
+    </Modal>
+  );
+};
+
+const styles = StyleSheet.create({
+  modal: {
+    justifyContent: "flex-end",
+    margin: 0,
+  },
+  container: {
+    borderTopLeftRadius: 20,
+    borderTopRightRadius: 20,
+    paddingHorizontal: 20,
+  },
+  handle: {
+    alignSelf: "center",
+    borderRadius: 2,
+    height: 4,
+    marginTop: 12,
+    width: 40,
+  },
+  header: {
+    alignItems: "center",
+    backgroundColor: "transparent",
+    flexDirection: "row",
+    justifyContent: "space-between",
+  },
+  title: {
+    fontSize: 20,
+  },
+});
+
+export default BottomSheetModal;
