+++ src/components/popups/ChangelogModal.tsx
@@ -1,9 +1,6 @@
-import { ScrollView, StyleSheet, View } from "react-native";
-import Modal from "react-native-modal";
-import { useSafeAreaInsets } from "react-native-safe-area-context";
+import { ScrollView, StyleSheet } from "react-native";
 
-import { CloseIcon } from "#/components/Icons";
-import Heading from "#/components/typography/Heading";
+import BottomSheetModal from "#/components/popups/BottomSheetModal";
 import UiPressable from "#/components/ui/UiPressable";
 import UiSpace from "#/components/ui/UiSpace";
 import UiText from "#/components/ui/UiText";
@@ -18,91 +15,42 @@ interface ChangelogModalProperties {
 
 const ChangelogModal = ({ isVisible, onClose }: ChangelogModalProperties) => {
   const colorScheme = useAppColorScheme();
-  const insets = useSafeAreaInsets();
   const corporate = Colors[colorScheme].primary;
-  const surface = Colors[colorScheme].surface;
   const textMuted = Colors[colorScheme].textMuted;
   const onPrimary = Colors[colorScheme].onPrimary;
 
   return (
-    <Modal
+    <BottomSheetModal
       isVisible={isVisible}
-      onSwipeComplete={onClose}
-      onBackdropPress={onClose}
-      swipeDirection="down"
-      style={styles.modal}
+      onClose={onClose}
+      title="Was ist neu?"
+      containerStyle={styles.container}
     >
-      <View
-        style={[
-          styles.container,
-          // Lift the sheet content (and the "Alles klar" button) above the
-          // Android nav bar / home indicator.
-          { backgroundColor: surface, paddingBottom: 20 + insets.bottom },
-        ]}
+      <UiText style={[styles.version, { color: textMuted }]}>
+        Version {Changelog.version}
+      </UiText>
+      <UiSpace size={12} />
+      <ScrollView style={styles.scroll} showsVerticalScrollIndicator={false}>
+        <UiText style={styles.notes}>{Changelog.notes}</UiText>
+      </ScrollView>
+      <UiSpace size={20} />
+      <UiPressable
+        accessibilityRole="button"
+        onPress={onClose}
+        style={[styles.button, { backgroundColor: corporate }]}
       >
-        <View style={[styles.handle, { backgroundColor: textMuted }]} />
-        <UiSpace size={16} />
-        <View style={styles.header}>
-          <Heading style={styles.title}>Was ist neu?</Heading>
-          <UiPressable
-            accessibilityRole="button"
-            accessibilityLabel="Schließen"
-            onPress={onClose}
-            hitSlop={8}
-          >
-            <CloseIcon size={28} color={corporate} />
-          </UiPressable>
-        </View>
-        <UiText style={[styles.version, { color: textMuted }]}>
-          Version {Changelog.version}
+        <UiText style={[styles.buttonText, { color: onPrimary }]}>
+          Alles klar
         </UiText>
-        <UiSpace size={12} />
-        <ScrollView style={styles.scroll} showsVerticalScrollIndicator={false}>
-          <UiText style={styles.notes}>{Changelog.notes}</UiText>
-        </ScrollView>
-        <UiSpace size={20} />
-        <UiPressable
-          accessibilityRole="button"
-          onPress={onClose}
-          style={[styles.button, { backgroundColor: corporate }]}
-        >
-          <UiText style={[styles.buttonText, { color: onPrimary }]}>
-            Alles klar
-          </UiText>
-        </UiPressable>
-        <UiSpace size={8} />
-      </View>
-    </Modal>
+      </UiPressable>
+      <UiSpace size={8} />
+    </BottomSheetModal>
   );
 };
 
 const styles = StyleSheet.create({
-  modal: {
-    justifyContent: "flex-end",
-    margin: 0,
-  },
   container: {
-    borderTopLeftRadius: 20,
-    borderTopRightRadius: 20,
-    paddingHorizontal: 20,
-    // paddingBottom is applied inline to include the safe-area inset.
     maxHeight: "80%",
-  },
-  handle: {
-    alignSelf: "center",
-    borderRadius: 2,
-    height: 4,
-    marginTop: 12,
-    width: 40,
-  },
-  header: {
-    alignItems: "center",
-    backgroundColor: "transparent",
-    flexDirection: "row",
-    justifyContent: "space-between",
-  },
-  title: {
-    fontSize: 20,
   },
   version: {
     fontSize: 13,
